copy load_from_tsfile_to_dataframe remove excessive whitespace
