+++ sktime/datasets/__init__.py
@@ -21,6 +21,7 @@ __all__ = [
     "load_electric_devices_segmentation",
     "load_acsf1",
     "load_macroeconomic",
+    "load_from_tsfile_to_dataframe",
 ]
 
 from sktime.datasets._data_io import (
@@ -29,6 +30,7 @@ from sktime.datasets._data_io import (
     load_arrow_head,
     load_basic_motions,
     load_electric_devices_segmentation,
+    load_from_tsfile_to_dataframe,
     load_gun_point_segmentation,
     load_gunpoint,
     load_italy_power_demand,
